// src/app/model/Workspace/ConsentsData.model.ts
export class ConsentsData
{
    HipaaViaMail: boolean;
    HipaaViaVoice: boolean;
    HipaaViaNotice: boolean;
    HipaaViaMessage: boolean;
    HipaaAllowSms: boolean;
    HipaaAllowEmail: boolean;
    HippaAllowPhoneCall: boolean;
    HippaAllowPatientPortal: boolean;
    AcceptAssignment: boolean;
    isApproved: boolean;
    isActive: boolean;
    PasswordPolicy: boolean;
}